fix: move LikeBtn away from Product
